chatbot-project/be/community_api/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
cryptography
mangum
boto3
websockets
redis[asyncio]
